requirements.txt
Django==3.2.2
requests==2.25.1
django-filter==2.4.0
djangorestframework==3.12.4
django-rest-auth==0.9.5
uritemplate==3.0.1
pyyaml==5.4.1
psycopg2-binary>=2.8
plotly==5.3.1
whitenoise==5.3.0
gunicorn==20.1.0
django-cors-headers==3.7.0
django-rest-passwordreset==1.2.1
Unidecode-1.3.4